Wrestling with OSX. Looks like I’ll have to do shader based rendering for everything or nothing.
Former-commit-id: 6513b9d4055f997418a0174fa006f40511d824e3

+++ Box2D/Testbed/Framework/RenderGL3.cpp
@@ -489,7 +489,7 @@ bool RenderGLInit(const char* fontpath)
 	FILE* fp = fopen(fontpath, "rb");
 	if (!fp) return false;
 	fseek(fp, 0, SEEK_END);
-	int size = ftell(fp);
+	int size = (int)ftell(fp);
 	fseek(fp, 0, SEEK_SET);
 
 	unsigned char* ttfBuffer = (unsigned char*)malloc(size);
